refactor: split Builtins.h to decouple lightweight consumers from LLVM headers
Builtins.h previously included both TypeChecker.h and IRGenerator.h in order
to form the method-pointer type aliases, which dragged the full LLVM header
tree into every TU that needed only builtin name constants or flag metadata.

Split into two headers:
- BuiltinFunctions.h: lightweight — name constants, BuiltinFunctionInfo
  (counts/flags only), BUILTIN_FUNCTIONS array, BUILTIN_FUNCTIONS_MAP, and
  the consteval validator. No TypeChecker or IRGenerator dependency.
- Builtins.h: thin dispatch shim — includes BuiltinFunctions.h plus both
  pass headers, adds TypeCheckerVisitMethod/IRGeneratorVisitMethod aliases,
  BuiltinFunctionDispatch struct, and BUILTIN_DISPATCH_MAP.

TypeCheckerBuiltinFunctions.cpp and GenBuiltinFunctions.cpp keep Builtins.h
and use BUILTIN_DISPATCH_MAP for the method-pointer lookup. The four
lightweight consumers (ASTNodes.cpp, TypeCheckerValues.cpp, GenValues.cpp,
PostTypeCheckingVerifier.cpp) now include only BuiltinFunctions.h.
